# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Sales & Distributor Commission Management API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export type Role = (typeof Role)[keyof typeof Role];

export const Role = {
  ADMIN: "ADMIN",
  DISTRIBUTOR: "DISTRIBUTOR",
  SALES: "SALES",
} as const;

export interface User {
  id: number;
  name: string;
  email: string;
  role: Role;
  /** @nullable */
  distributorId?: number | null;
  createdAt: string;
}

export interface LoginRequest {
  email: string;
  password: string;
}

export interface AuthSession {
  token: string;
  user: User;
}

export interface CreateUserRequest {
  name: string;
  email: string;
  password: string;
  role: Role;
  /** @nullable */
  distributorId?: number | null;
}

export interface Client {
  id: number;
  name: string;
  taxCardNumber: string;
  taxCardName: string;
  issuingAuthority: string;
  commercialRegistryNumber: string;
  businessType: string;
  email: string;
  phone1: string;
  phone1WhatsApp: boolean;
  /** @nullable */
  phone2?: string | null;
  phone2WhatsApp?: boolean;
  nationalId: string;
  address: string;
  assignedSalesId: number;
  assignedDistributorId: number;
  ownershipStartDate: string;
  ownershipEndDate: string;
  createdAt: string;
}

export interface UpdateClientRequest {
  name?: string;
  taxCardNumber?: string;
  taxCardName?: string;
  issuingAuthority?: string;
  commercialRegistryNumber?: string;
  businessType?: string;
  email?: string;
  phone1?: string;
  phone1WhatsApp?: boolean;
  /** @nullable */
  phone2?: string | null;
  phone2WhatsApp?: boolean;
  nationalId?: string;
  address?: string;
}

export interface CreateClientRequest {
  name: string;
  taxCardNumber: string;
  taxCardName: string;
  issuingAuthority: string;
  commercialRegistryNumber: string;
  businessType: string;
  email: string;
  phone1: string;
  phone1WhatsApp: boolean;
  /** @nullable */
  phone2?: string | null;
  phone2WhatsApp?: boolean;
  nationalId: string;
  address: string;
  /** @nullable */
  assignedSalesId?: number | null;
}

export interface ClientLookupResult {
  found: boolean;
  client?: Client;
  hasPendingOrder?: boolean;
}

export type OrderStatus = (typeof OrderStatus)[keyof typeof OrderStatus];

export const OrderStatus = {
  PENDING: "PENDING",
  COMPLETED: "COMPLETED",
} as const;

export interface Order {
  id: number;
  clientId: number;
  /** @nullable */
  clientName?: string | null;
  /** @nullable */
  packageId?: number | null;
  /** @nullable */
  packageName?: string | null;
  /** @nullable */
  salesId?: number | null;
  /** @nullable */
  salesName?: string | null;
  /** @nullable */
  distributorId?: number | null;
  /** @nullable */
  distributorName?: string | null;
  orderName: string;
  amount: number;
  vatAmount: number;
  /** @nullable */
  receiptNumber?: string | null;
  isFullyCollected: boolean;
  orderDate: string;
  status: OrderStatus;
  createdAt: string;
}

export interface Client360Profile {
  client: Client;
  orders: Order[];
}

export interface ReassignClientRequest {
  assignedSalesId: number;
}

export interface ClientAssignment {
  id: number;
  clientId: number;
  /** @nullable */
  fromSalesId?: number | null;
  /** @nullable */
  fromSalesName?: string | null;
  /** @nullable */
  fromDistributorId?: number | null;
  /** @nullable */
  fromDistributorName?: string | null;
  toSalesId: number;
  /** @nullable */
  toSalesName?: string | null;
  toDistributorId: number;
  /** @nullable */
  toDistributorName?: string | null;
  changedById: number;
  /** @nullable */
  changedByName?: string | null;
  createdAt: string;
}

export interface Package {
  id: number;
  name: string;
  price: number;
  vatPct: number;
  isActive: boolean;
  createdAt: string;
}

export interface CreatePackageRequest {
  name: string;
  price: number;
  vatPct: number;
}

export interface UpdatePackageRequest {
  name?: string;
  price?: number;
  vatPct?: number;
  isActive?: boolean;
}

export interface CreateOrderRequest {
  clientId: number;
  packageId: number;
  /** @nullable */
  receiptNumber?: string | null;
  isFullyCollected: boolean;
}

export interface UpdateOrderStatusRequest {
  status: OrderStatus;
}

export type CommissionStatus =
  (typeof CommissionStatus)[keyof typeof CommissionStatus];

export const CommissionStatus = {
  UNPAID: "UNPAID",
  PAID: "PAID",
} as const;

export interface UpdateCommissionStatusRequest {
  status: CommissionStatus;
}

export interface MarkCommissionsPaidRequest {
  ids: number[];
}

export type CommissionRoleType =
  (typeof CommissionRoleType)[keyof typeof CommissionRoleType];

export const CommissionRoleType = {
  DISTRIBUTOR: "DISTRIBUTOR",
  SALES: "SALES",
} as const;

export type CommissionStatusProperty =
  (typeof CommissionStatusProperty)[keyof typeof CommissionStatusProperty];

export const CommissionStatusProperty = {
  UNPAID: "UNPAID",
  PAID: "PAID",
} as const;

export interface Commission {
  id: number;
  orderId: number;
  /** @nullable */
  orderName?: string | null;
  /** @nullable */
  clientName?: string | null;
  userId: number;
  /** @nullable */
  userName?: string | null;
  amount: number;
  roleType: CommissionRoleType;
  status: CommissionStatusProperty;
  createdAt: string;
  /** @nullable */
  paidAt?: string | null;
  /** @nullable */
  paidByUserId?: number | null;
  /** @nullable */
  paidByName?: string | null;
}

export interface DashboardSummary {
  role: Role;
  totalOrders: number;
  completedOrders: number;
  pendingOrders: number;
  totalSales: number;
  totalCommissions: number;
  /** @nullable */
  teamSize?: number | null;
  /** @nullable */
  clientCount?: number | null;
  recentOrders: Order[];
  recentCommissions: Commission[];
}

export interface CommissionRates {
  salesPct: number;
  distributorPct: number;
}

export type ListUsersParams = {
  role?: ListUsersRole;
  distributorId?: number;
};

export type ListUsersRole = (typeof ListUsersRole)[keyof typeof ListUsersRole];

export const ListUsersRole = {
  ADMIN: "ADMIN",
  DISTRIBUTOR: "DISTRIBUTOR",
  SALES: "SALES",
} as const;

export type LookupClientParams = {
  taxCardNumber: string;
};