# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Sales & Distributor Commission Management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /clients:
    get:
      operationId: listClients
      tags: [clients]
      responses:
        "200":
          description: Clients
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Client" }
    post:
      operationId: createClient
      tags: [clients]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateClientRequest" }
      responses:
        "201":
          description: Client created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Client" }
  /clients/lookup:
    get:
      operationId: lookupClient
      tags: [clients]
      parameters:
        - name: taxCardNumber
          in: query
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Client lookup result
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ClientLookupResult" }
  /clients/{id}:
    get:
      operationId: getClientProfile
      tags: [clients]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Client Profile
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Client360Profile" }
        "404":
          description: Client not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
    patch:
      operationId: updateClient
      tags: [clients]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateClientRequest" }
      responses:
        "200":
          description: Client updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Client" }
        "404":
          description: Client not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
  /clients/{id}/assignment:
    patch:
      operationId: reassignClient
      tags: [clients]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ReassignClientRequest" }
      responses:
        "200":
          description: Client reassigned
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Client" }
        "400":
          description: Bad request
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "403":
          description: Forbidden
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "404":
          description: Client not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
  /clients/{id}/assignments:
    get:
      operationId: listClientAssignments
      tags: [clients]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Assignment history for client
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/ClientAssignment" }
        "403":
          description: Forbidden
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "404":
          description: Client not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error: { type: string }
...
    Client:
      type: object
      required: [
        id, name, taxCardNumber, taxCardName, issuingAuthority, 
        commercialRegistryNumber, businessType, email, phone1, 
        phone1WhatsApp, nationalId, address, assignedSalesId, 
        assignedDistributorId, ownershipStartDate, ownershipEndDate, createdAt
      ]
      properties:
        id: { type: integer }
        name: { type: string }
        taxCardNumber: { type: string }
        taxCardName: { type: string }
        issuingAuthority: { type: string }
        commercialRegistryNumber: { type: string }
        businessType: { type: string }
        email: { type: string }
        phone1: { type: string }
        phone1WhatsApp: { type: boolean }
        phone2: { type: ["string", "null"] }
        phone2WhatsApp: { type: boolean }
        nationalId: { type: string }
        address: { type: string }
        assignedSalesId: { type: integer }
        assignedDistributorId: { type: integer }
        ownershipStartDate: { type: string, format: date-time }
        ownershipEndDate: { type: string, format: date-time }
        createdAt: { type: string, format: date-time }
    UpdateClientRequest:
      type: object
      properties:
        name: { type: string }
        taxCardNumber: { type: string }
        taxCardName: { type: string }
        issuingAuthority: { type: string }
        commercialRegistryNumber: { type: string }
        businessType: { type: string }
        email: { type: string }
        phone1: { type: string }
        phone1WhatsApp: { type: boolean }
        phone2: { type: ["string", "null"] }
        phone2WhatsApp: { type: boolean }
        nationalId: { type: string }
        address: { type: string }
    CreateClientRequest:
      type: object
      required: [
        name, taxCardNumber, taxCardName, issuingAuthority, 
        commercialRegistryNumber, businessType, email, phone1, 
        phone1WhatsApp, nationalId, address
      ]
      properties:
        name: { type: string }
        taxCardNumber: { type: string }
        taxCardName: { type: string }
        issuingAuthority: { type: string }
        commercialRegistryNumber: { type: string }
        businessType: { type: string }
        email: { type: string }
        phone1: { type: string }
        phone1WhatsApp: { type: boolean }
        phone2: { type: ["string", "null"] }
        phone2WhatsApp: { type: boolean }
        nationalId: { type: string }
        address: { type: string }
        assignedSalesId: { type: ["integer", "null"] }
    ClientLookupResult:
      type: object
      required: [found]
      properties:
        found: { type: boolean }
        client: { $ref: "#/components/schemas/Client" }
        hasPendingOrder: { type: boolean }
    Client360Profile:
      type: object
      required: [client, orders]
      properties:
        client: { $ref: "#/components/schemas/Client" }
        orders: 
          type: array
          items: { $ref: "#/components/schemas/Order" }
    ReassignClientRequest:
      type: object
      required: [assignedSalesId]
      properties:
        assignedSalesId: { type: integer }
    ClientAssignment:
      type: object
      required: [id, clientId, toSalesId, toDistributorId, changedById, createdAt]
      properties:
        id: { type: integer }
        clientId: { type: integer }
        fromSalesId: { type: ["integer", "null"] }
        fromSalesName: { type: ["string", "null"] }
        fromDistributorId: { type: ["integer", "null"] }
        fromDistributorName: { type: ["string", "null"] }
        toSalesId: { type: integer }
        toSalesName: { type: ["string", "null"] }
        toDistributorId: { type: integer }
        toDistributorName: { type: ["string", "null"] }
        changedById: { type: integer }
        changedByName: { type: ["string", "null"] }
        createdAt: { type: string, format: date-time }
...
    OrderStatus:
      type: string
      enum: [PENDING, COMPLETED]
    Order:
      type: object
      required: [id, clientId, orderName, amount, vatAmount, isFullyCollected, orderDate, status, createdAt]
      properties:
        id: { type: integer }
        clientId: { type: integer }
        clientName: { type: ["string", "null"] }
        packageId: { type: ["integer", "null"] }
        packageName: { type: ["string", "null"] }
        salesId: { type: ["integer", "null"] }
        salesName: { type: ["string", "null"] }
        distributorId: { type: ["integer", "null"] }
        distributorName: { type: ["string", "null"] }
        orderName: { type: string }
        amount: { type: number }
        vatAmount: { type: number }
        receiptNumber: { type: ["string", "null"] }
        isFullyCollected: { type: boolean }
        orderDate: { type: string, format: date-time }
        status: { $ref: "#/components/schemas/OrderStatus" }
        createdAt: { type: string, format: date-time }
...