# lib/api-zod/src/generated/types/client.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Sales & Distributor Commission Management API
 * OpenAPI spec version: 0.1.0
 */

export interface Client {
  id: number;
  name: string;
  taxCardNumber: string;
  taxCardName: string;
  issuingAuthority: string;
  commercialRegistryNumber: string;
  businessType: string;
  email: string;
  phone1: string;
  phone1WhatsApp: boolean;
  /** @nullable */
  phone2?: string | null;
  phone2WhatsApp?: boolean;
  nationalId: string;
  address: string;
  assignedSalesId: number;
  assignedDistributorId: number;
  ownershipStartDate: Date;
  ownershipEndDate: Date;
  createdAt: Date;
}