# lib/api-zod/src/generated/types/client360Profile.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Sales & Distributor Commission Management API
 * OpenAPI spec version: 0.1.0
 */
import type { Client } from "./client";
import type { Order } from "./order";

export interface Client360Profile {
  client: Client;
  orders: Order[];
}