# lib/api-zod/src/generated/types/clientLookupResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Sales & Distributor Commission Management API
 * OpenAPI spec version: 0.1.0
 */
import type { Client } from "./client";

export interface ClientLookupResult {
  found: boolean;
  client?: Client;
  hasPendingOrder?: boolean;
}