# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Sales & Distributor Commission Management API
 * OpenAPI spec version: 0.1.0
 */
import type { OrderStatus } from "./orderStatus";

export interface Order {
  id: number;
  clientId: number;
  /** @nullable */
  clientName?: string | null;
  /** @nullable */
  packageId?: number | null;
  /** @nullable */
  packageName?: string | null;
  /** @nullable */
  salesId?: number | null;
  /** @nullable */
  salesName?: string | null;
  /** @nullable */
  distributorId?: number | null;
  /** @nullable */
  distributorName?: string | null;
  orderName: string;
  amount: number;
  vatAmount: number;
  /** @nullable */
  receiptNumber?: string | null;
  isFullyCollected: boolean;
  orderDate: Date;
  status: OrderStatus;
  createdAt: Date;
}